requirements.txt
beautifulsoup4==4.6.0
blueware-enterprise==1.2.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
Django==1.11
django-cors-headers==2.3.0
django-filter==1.1.0
django-filters==0.2.1
django-qiniu-storage==2.3.1
djangorestframework==3.7.7
httpie==0.9.9
idna==2.6
jsonpickle==0.9.6
lxml==4.1.1
Pillow==5.0.0
Pygments==2.2.0
pytz==2017.3
qiniu==7.2.2
requests==2.18.4
six==1.11.0
urllib3==1.22
